requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.10
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.1
certifi==2022.9.14
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
easyocr==1.6.2
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.2
filelock==3.8.0
identify==2.5.5
idna==3.4
imageio==2.22.1
imutils==0.5.4
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mypy-extensions==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
networkx==2.8.7
ninja==1.10.2.4
nodeenv==1.7.0
notebook==6.4.12
numpy==1.23.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.0
pure-eval==0.2.2
pyclipper==1.3.0.post3
pycparser==2.21
Pygments==2.13.0
pylint==2.15.3
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
pywin32==303; platform_system == "Windows"
pywinpty==2.0.8; platform_system == "Windows"
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.2
Send2Trash==1.8.0
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tensorflow==2.10.0
terminado==0.15.0
tifffile==2022.8.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch==1.12.1
torchvision==0.13.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
wincertstore==0.2
wrapt==1.14.1
-e .
